Reorganize into one pldp app
