--- pldp/study/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class StudyConfig(AppConfig):
-    name = 'study'
